updated form and table in well card

+++ resources/js/components/bigdata/forms/PlainFormResults.vue
@@ -1,73 +1,75 @@
 <template>
-  <div>
-    <div v-if="rows" class="table-container">
-      <div class="table-container-header">
+  <div v-if="rows" class="table-container">
+    <div class="table-container-header">
+
+      <svg fill="none" height="16" viewBox="0 0 16 16" width="16" xmlns="http://www.w3.org/2000/svg"
+           @click="showForm()">
+        <path d="M14.5 8L1.5 8" stroke="white" stroke-linecap="round" stroke-width="1.5"/>
+        <path d="M8 1.5V14.5" stroke="white" stroke-linecap="round" stroke-width="1.5"/>
+      </svg>
+      <template v-if="header">
         <div class="row">
           <div class="col">
-            <svg fill="none" height="16" viewBox="0 0 16 16" width="16" xmlns="http://www.w3.org/2000/svg"
-                 @click="showForm()">
-              <path d="M14.5 8L1.5 8" stroke="white" stroke-linecap="round" stroke-width="1.5"/>
-              <path d="M8 1.5V14.5" stroke="white" stroke-linecap="round" stroke-width="1.5"/>
+            <h4>
+              Текущая оргструктура: <span
+                class="blue-section">НИИ/АО “РД “КазМунайГаз”/АО “Озенмунайгаз”/НГДУ-З/ППД-З</span>
+            </h4>
+          </div>
+        </div>
+        <div class="row">
+          <div class="col">
+            <h4>
+              Начало периода: 01.01.1980
+            </h4>
+            <h4>
+              Конец периода: ...
+            </h4>
+          </div>
+        </div>
+      </template>
+    </div>
+    <table class="table-container-table">
+      <thead class="table-container-column-header">
+      <tr>
+        <th class="table-border"><p class="title">ID</p></th>
+        <th v-for="column in columns" class="table-border"><p class="title">{{ column.title }}</p>
+        </th>
+        <th class="table-border"><p class="title">Управление</p></th>
+      </tr>
+      </thead>
+      <tbody class="table-container-element">
+      <tr v-for="(row, index) in rows">
+        <td class="table-border element-position">
+          <p class="title">{{ row.id }}</p>
+        </td>
+        <td v-for="column in columns" class="table-border element-position">
+          <p v-if="typeof dictFields[column.code] !== 'undefined' && typeof getDict(dictFields[column.code])[row[column.code]] !== 'undefined'">
+            {{ getDict(dictFields[column.code])[row[column.code]].name }}
+          </p>
+          <p v-else>
+            {{ row[column.code] }}
+          </p>
+        </td>
+        <td class="table-border element-position">
+          <div class="table-container-svg">
+            <svg fill="none" height="18" viewBox="0 0 18 18" width="18" xmlns="http://www.w3.org/2000/svg"
+                 @click.prevent="editRow(row)">
+              <path
+                  d="M3 11.4998L1.55336 16.322C1.53048 16.3983 1.6016 16.4694 1.67788 16.4465L6.5 14.9998M3 11.4998C3 11.4998 11.0603 3.4393 12.7227 1.77708C12.8789 1.62091 13.1257 1.6256 13.2819 1.78177C13.8372 2.33702 15.1144 3.61422 16.2171 4.71697C16.3733 4.87322 16.3788 5.12103 16.2226 5.27726C14.5597 6.9399 6.5 14.9998 6.5 14.9998M3 11.4998L3.64727 10.8525L7.14727 14.3525L6.5 14.9998"
+                  stroke="white" stroke-width="1.4"/>
+            </svg>
+            <svg fill="none" height="14" viewBox="0 0 14 14" width="14" xmlns="http://www.w3.org/2000/svg"
+                 @click.prevent="deleteRow(row, index)">
+              <path d="M12.6574 12.6575L1.34367 1.34383" stroke="white" stroke-linecap="round"
+                    stroke-width="1.4"/>
+              <path d="M12.6563 1.34383L1.34262 12.6575" stroke="white" stroke-linecap="round"
+                    stroke-width="1.4"/>
             </svg>
           </div>
-        </div>
-        <template v-if="header">
-          <div class="row">
-            <div class="col">
-              <h4>
-                Текущая оргструктура: <span
-                  class="blue-section">НИИ/АО “РД “КазМунайГаз”/АО “Озенмунайгаз”/НГДУ-З/ППД-З</span>
-              </h4>
-            </div>
-          </div>
-          <div class="row">
-            <div class="col">
-              <h4>
-                Начало периода: 01.01.1980
-              </h4>
-              <h4>
-                Конец периода: ...
-              </h4>
-            </div>
-          </div>
-        </template>
-      </div>
-      <div class="table-container-column-header">
-        <div class="row">
-          <div class="col-md-2 table-border flex"><p class="title">ID</p></div>
-          <div v-for="column in columns" class="col-md-2 table-border flex"><p class="title">{{ column.title }}</p>
-          </div>
-          <div class="col-md-2 table-border flex"><p class="title">Управление</p></div>
-        </div>
-      </div>
-      <div class="table-container-element">
-        <div v-for="(row, index) in rows" class="row">
-          <div class="col-md-2 table-border element-position">
-            <p class="title">{{ row.id }}</p>
-          </div>
-          <div v-for="column in columns" class="col-md-2 table-border element-position"><p>{{ row[column.code] }}</p>
-          </div>
-          <div class="col-md-2 table-border element-position"><p>7.21</p></div>
-          <div class="col-md-2 table-border svg-element">
-            <div class="table-container-svg">
-              <svg fill="none" height="18" viewBox="0 0 18 18" width="18" xmlns="http://www.w3.org/2000/svg"
-                   @click.prevent="editRow(row)">
-                <path
-                    d="M3 11.4998L1.55336 16.322C1.53048 16.3983 1.6016 16.4694 1.67788 16.4465L6.5 14.9998M3 11.4998C3 11.4998 11.0603 3.4393 12.7227 1.77708C12.8789 1.62091 13.1257 1.6256 13.2819 1.78177C13.8372 2.33702 15.1144 3.61422 16.2171 4.71697C16.3733 4.87322 16.3788 5.12103 16.2226 5.27726C14.5597 6.9399 6.5 14.9998 6.5 14.9998M3 11.4998L3.64727 10.8525L7.14727 14.3525L6.5 14.9998"
-                    stroke="white" stroke-width="1.4"/>
-              </svg>
-              <svg fill="none" height="14" viewBox="0 0 14 14" width="14" xmlns="http://www.w3.org/2000/svg"
-                   @click.prevent="deleteRow(row, index)">
-                <path d="M12.6574 12.6575L1.34367 1.34383" stroke="white" stroke-linecap="round"
-                      stroke-width="1.4"/>
-                <path d="M12.6563 1.34383L1.34262 12.6575" stroke="white" stroke-linecap="round"
-                      stroke-width="1.4"/>
-              </svg>
-            </div>
-          </div>
-        </div>
-      </div>
-    </div>
+        </td>
+      </tr>
+      </tbody>
+    </table>
     <div v-if="isFormOpened" class="bd-popup">
       <div class="bd-popup__inner">
         <a class="bd-popup__close" href="#" @click.prevent="isFormOpened = false">{{ trans('bd.close') }}</a>
@@ -76,6 +78,8 @@
               :params="formParams"
               :values="formValues"
               :well-id="wellId"
+              @change="updateResults"
+              @close="isFormOpened = false"
           >
           </BigDataPlainForm>
         </div>
@@ -111,19 +115,42 @@ export default {
       columns: null,
       isFormOpened: false,
       formValues: null,
-      formParams: null
+      formParams: null,
+      dictFields: {}
     }
   },
   watch: {
     code() {
       this.updateResults()
-    }
+    },
+    wellId() {
+      this.updateResults()
+    },
   },
   mounted() {
     this.updateResults()
   },
   methods: {
     updateResults() {
+
+      this.axios.get(this.localeUrl(`/api/bigdata/forms/${this.code}`)).then(({data}) => {
+
+        data.params.tabs.forEach(tab => {
+          tab.blocks.forEach(blocks => {
+            blocks.forEach(block => {
+
+              block.items
+                  .filter(item => item.type === 'dict')
+                  .map(item => {
+                    this.dictFields[item.code] = item.dict
+                  })
+
+            })
+          })
+        })
+        this.loadDictionaries()
+      })
+
       this.axios.get(
           this.localeUrl(`/api/bigdata/forms/${this.code}/results`),
           {params: {well_id: this.wellId}}
@@ -132,17 +159,41 @@ export default {
         this.columns = data.columns
       })
     },
+    loadDictionaries() {
+      Object.values(this.dictFields).forEach(code => {
+        if (this.getDict(code)) return
+        this.axios.get(this.localeUrl(`/api/bigdata/dict/${code}`)).then(data => {
+          this.$store.commit("bd/SAVE_DICT", {
+            code: code,
+            items: data.data
+          });
+        })
+      })
+    },
+    getDict(code) {
+      return this.$store.getters['bd/dict'](code);
+    },
     showForm() {
       this.formParams = this.forms.find(form => form.code === this.code)
       this.isFormOpened = true
     },
     editRow(row) {
-      this.$emit('edit', row)
+      this.formParams = this.forms.find(form => form.code === this.code)
+      this.formValues = row
+      this.isFormOpened = true
     },
     deleteRow(row, rowIndex) {
-      this.axios.delete(this.localeUrl(`/api/bigdata/forms/${this.code}/${row.id}`)).then(({data}) => {
-        this.rows.splice(rowIndex, 1)
+      this.$bvModal.msgBoxConfirm('Вы хотите удалить запись?', {
+        okTitle: this.trans('app.yes'),
+        cancelTitle: this.trans('app.no'),
       })
+          .then(result => {
+            if (result === true) {
+              this.axios.delete(this.localeUrl(`/api/bigdata/forms/${this.code}/${row.id}`)).then(({data}) => {
+                this.rows.splice(rowIndex, 1)
+              })
+            }
+          })
     }
   }
 }
@@ -155,16 +206,34 @@ export default {
   width: 100%;
   color: white;
 
-  .table-container-header {
-    text-align: center;
-    padding: 14px 20px 0px 20px;
+  &-table {
+    width: 100%;
+
+    th {
+      font-weight: normal;
+      padding: 5px 0;
+      vertical-align: middle;
+
+      p {
+        margin: 0;
+      }
+    }
+
+    tr:nth-child(2n) {
+      td {
+        background: #31355E;
+      }
+    }
+  }
+
+  &-header {
+    text-align: right;
+    padding: 14px 20px;
     background-color: #32346C;
   }
 
-  .table-container-column-header {
-    align-items: center;
+  &-column-header {
     background-color: #505684;
-    display: flex;
     min-height: 50px;
     text-align: center;
 
@@ -175,7 +244,6 @@ export default {
   }
 
   .table-container-element {
-    height: 340px;
     background-color: #272953;
 
     .table-container-svg {
@@ -195,8 +263,7 @@ export default {
     }
 
     .element-position {
-      padding: 5px 13px 5px 23px;
-      display: flex;
+      padding: 9px 13px;
 
       p {
         float: right;
@@ -225,22 +292,14 @@ export default {
   }
 
   .row {
-    margin-right: 0px;
+    margin-right: 0;
   }
 
   .table-border {
-    align-items: center;
-    border-top: hidden;
-    border-left: 1px solid #464D7A;
-    border-bottom: hidden;
-    border-right: hidden;
-    display: flex;
-    justify-content: center;
-
-    p {
-      margin-top: auto;
-      margin-bottom: auto;
-    }
+    border: 1px solid #464D7A;
+    border-bottom: none;
+    border-top: none;
+    vertical-align: middle;
   }
 }
 </style>
